third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/switch-panels-while-editing-as-html-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verifies inspector doesn't break when switching panels while editing as HTML. crbug.com/485457


Running: testSetUp

Running: selectNode

Running: switchPanels

Running: switchBackToElements
Is editing: false
